build and inspect
this commits extends the UI with the following flag:

- `--bin`
- `--example`
- `--lib`
- `--release`

if one of `--bin`, `--example` or `--lib` is used then cargo-$tool will first
build the Cargo project and then proceed to run the llvm tool on the output
artifact -- that is the output artifact will be the first argument passed to the
llvm tool.

`--release` controls whether the project is compiled with optimizations or not.

if none of these three flags is passed then the old behavior is used. No path to
an artifact is passed to the llvm tool and the user is expected to provide it.

this new logic supports workspaces and the CARGO_TARGET_DIR env variable.

closes #15