Do not treat features as mutually exclusive
It's possible to use '--no-default-features --features' foo to build
a Rust crate. But, we were treating these flags as mutually
exclusive. The commit treats the flags as not exclusive, satisfying
the following usage:

  cargo objcopy --no-default-features --features foo

Previously, if run with the `--verbose` flag, we would notice that
'--no-default-features` would not be printed. Now, it is included.